request join team and validate join request team

+++ app/Http/Controllers/TeamController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\MasterTeam;
+use App\Models\Personnel;
+use App\Models\MasterReqJoinTeam;
 use Illuminate\Support\Facades\DB;
 use stdClass;
 use Illuminate\Support\Facades\Validator;
@@ -24,7 +26,7 @@ class TeamController extends Controller
         $response->desc = '';
         $requestData = $request->input();
         DB::beginTransaction();
-        try {;
+        try {
             $validator = Validator::make($requestData, [
                 'user_id' => 'required|string',
                 'name' => 'required|string',
@@ -36,20 +38,33 @@ class TeamController extends Controller
             $teamInfo = isset($requestData['info']) ? trim($requestData['info']) : NULL;
             $teamPersonnel = isset($requestData['personnel']) ? json_decode($requestData['personnel']) : NULL;
             if (!$validator->fails()) {
-                $checkTeamName = MasterTeam::where('name', $teamName)->first();
-                if (!$checkTeamName) {
-                    $insertData = array(
-                        'name' => $teamName,
-                        'info' => $teamInfo,
-                        'admin_id' => $adminId,
-                        'personnel' => json_encode($teamPersonnel),
-                    );
-                    MasterTeam::create($insertData);
-                    $response->code = '00';
-                    $response->desc = 'Create Team Success!';
+                $checkPersonnelRole = Personnel::where('user_id', $adminId)
+                    ->where('role', '2')
+                    ->first();
+                if ($checkPersonnelRole) {
+                    $checkTeamName = MasterTeam::where('name', $teamName)->first();
+                    if (!$checkTeamName) {
+                        $insertData = array(
+                            'name' => $teamName,
+                            'info' => $teamInfo,
+                            'admin_id' => $adminId,
+                            'personnel' => json_encode($teamPersonnel),
+                        );
+                        $createTeam = MasterTeam::create($insertData);
+                        $updatePersonnelTeam = array(
+                            'team_id' => $createTeam->id
+                        );
+                        Personnel::where('user_id', $adminId)
+                            ->update($updatePersonnelTeam);
+                        $response->code = '00';
+                        $response->desc = 'Create Team Success!';
+                    } else {
+                        $response->code = '02';
+                        $response->desc = 'Team Name Already Exist.';
+                    }
                 } else {
                     $response->code = '02';
-                    $response->desc = 'Team Name Already Exist.';
+                    $response->desc = 'You not have access.';
                 }
             } else {
                 $response->code = '01';
@@ -71,7 +86,7 @@ class TeamController extends Controller
         $response->desc = '';
         $requestData = $request->input();
         DB::beginTransaction();
-        try {;
+        try {
             $validator = Validator::make($requestData, [
                 'user_id' => 'required|string',
                 'team_id' => 'required|string',
@@ -85,20 +100,28 @@ class TeamController extends Controller
             $teamInfo = isset($requestData['info']) ? trim($requestData['info']) : NULL;
             $teamPersonnel = isset($requestData['personnel']) ? json_decode($requestData['personnel']) : NULL;
             if (!$validator->fails()) {
-                $checkTeamExist = MasterTeam::where('id', $teamId)->first();
-                if ($checkTeamExist) {
-                    $updateData = array(
-                        'name' => $teamName,
-                        'info' => $teamInfo,
-                        'admin_id' => $adminId,
-                        'personnel' => json_encode($teamPersonnel),
-                    );
-                    MasterTeam::where('id', $teamId)->update($updateData);
-                    $response->code = '00';
-                    $response->desc = 'Update Team Success!';
+                $checkPersonnelRole = Personnel::where('user_id', $adminId)
+                    ->where('role', '2')
+                    ->first();
+                if ($checkPersonnelRole) {
+                    $checkTeamExist = MasterTeam::where('id', $teamId)->first();
+                    if ($checkTeamExist) {
+                        $updateData = array(
+                            'name' => $teamName,
+                            'info' => $teamInfo,
+                            'admin_id' => $adminId,
+                            'personnel' => json_encode($teamPersonnel),
+                        );
+                        MasterTeam::where('id', $teamId)->update($updateData);
+                        $response->code = '00';
+                        $response->desc = 'Update Team Success!';
+                    } else {
+                        $response->code = '02';
+                        $response->desc = 'Team Not Found.';
+                    }
                 } else {
                     $response->code = '02';
-                    $response->desc = 'Team Not Found.';
+                    $response->desc = 'You not have access.';
                 }
             } else {
                 $response->code = '01';
@@ -120,7 +143,7 @@ class TeamController extends Controller
         $response->desc = '';
         $requestData = $request->input();
         DB::beginTransaction();
-        try {;
+        try {
             $validator = Validator::make($requestData, [
                 'user_id' => 'required|string',
                 'team_id' => 'required|string',
@@ -128,14 +151,27 @@ class TeamController extends Controller
             $adminId = isset($requestData['user_id']) ? trim($requestData['user_id']) : NULL;
             $teamId = isset($requestData['team_id']) ? trim($requestData['team_id']) : NULL;
             if (!$validator->fails()) {
-                $checkTeamExist = MasterTeam::where('id', $teamId)->first();
-                if ($checkTeamExist) {
-                    MasterTeam::where('id', $teamId)->delete();
-                    $response->code = '00';
-                    $response->desc = 'Delete Team Success!';
+                $checkPersonnelRole = Personnel::where('user_id', $adminId)
+                    ->where('role', '2')
+                    ->first();
+                if ($checkPersonnelRole) {
+                    $checkTeamExist = MasterTeam::where('id', $teamId)->first();
+                    if ($checkTeamExist) {
+                        MasterTeam::where('id', $teamId)->delete();
+                        $updatePersonnelTeam = array(
+                            'team_id' => NULL
+                        );
+                        Personnel::where('user_id', $adminId)
+                            ->update($updatePersonnelTeam);
+                        $response->code = '00';
+                        $response->desc = 'Delete Team Success!';
+                    } else {
+                        $response->code = '02';
+                        $response->desc = 'Team Not Found.';
+                    }
                 } else {
                     $response->code = '02';
-                    $response->desc = 'Team Not Found.';
+                    $response->desc = 'You not have access.';
                 }
             } else {
                 $response->code = '01';
@@ -157,7 +193,7 @@ class TeamController extends Controller
         $response->desc = '';
         $requestData = $request->input();
         DB::beginTransaction();
-        try {;
+        try {
             $validator = Validator::make($requestData, [
                 'user_id' => 'required|string',
             ]);
@@ -192,7 +228,7 @@ class TeamController extends Controller
         $response->desc = '';
         $requestData = $request->input();
         DB::beginTransaction();
-        try {;
+        try {
             $validator = Validator::make($requestData, [
                 'user_id' => 'required|string',
                 'team_id' => 'required|string',
@@ -205,6 +241,63 @@ class TeamController extends Controller
                     $response->code = '00';
                     $response->desc = 'Get Info Team Success!';
                     $response->data = $getInfoTeam;
+                } else {
+                    $response->code = '02';
+                    $response->desc = 'Team Not Found.';
+                }
+            } else {
+                $response->code = '01';
+                $response->desc = $validator->errors()->first();
+            }
+            DB::commit();
+        } catch (Exception $e) {
+            DB::rollback();
+            $response->responseCode = '99';
+            $response->responseDesc = 'Caught exception: ' .  $e->getMessage();
+        }
+        return response()->json($response);
+    }
+
+    public function answerReqJoinTeam(Request $request)
+    {
+        $response = new stdClass();
+        $response->code = '';
+        $response->desc = '';
+        $requestData = $request->input();
+        DB::beginTransaction();
+        try {
+            $validator = Validator::make($requestData, [
+                'user_id' => 'required|numeric',
+                'user_id_requested' => 'required|numeric',
+                'answer' => 'required|numeric',
+            ]);
+            $adminId = trim($requestData['user_id']);
+            $user_id_requested = trim($requestData['user_id_requested']);
+            $answer = trim($requestData['answer']);
+            if (!$validator->fails()) {
+                $checkTeamExist = MasterTeam::where('admin_id', $adminId)->first();
+                if ($checkTeamExist) {
+                    if ($answer == 1) {
+                        $arrayPersonnelTeam = array();
+                        if (is_array($checkTeamExist['personnel'])) {
+                            $arrayPersonnelTeam = $checkTeamExist['personnel'];
+                        }
+                        array_push($arrayPersonnelTeam, $user_id_requested);
+                        MasterTeam::where('id', $checkTeamExist['id'])
+                            ->update(array(
+                                'personnel' => json_encode($arrayPersonnelTeam)
+                            ));
+                        Personnel::where('user_id', $user_id_requested)
+                            ->update(array(
+                                'team_id' => $checkTeamExist['id']
+                            ));
+                        MasterReqJoinTeam::where('user_id', $user_id_requested)->delete();
+                    } else {
+                        MasterReqJoinTeam::where('user_id', $user_id_requested)
+                            ->update(array('answer' => 0));
+                    }
+                    $response->code = '00';
+                    $response->desc = 'Answer Request Success!';
                 } else {
                     $response->code = '02';
                     $response->desc = 'Team Not Found.';
